change dataType id in logisticaddaddress

+++ features/logistic/logisticaddaddress/src/main/java/com/tokopedia/logisticaddaddress/domain/model/Address.java
@@ -14,11 +14,11 @@ import java.util.ArrayList;
 
 public class Address implements Parcelable, Visitable<DistrictTypeFactory> {
 
-    private int districtId;
+    private long districtId;
     private String districtName;
-    private int cityId;
+    private long cityId;
     private String cityName;
-    private int provinceId;
+    private long provinceId;
     private String provinceName;
     private ArrayList<String> zipCodes;
 
@@ -26,11 +26,11 @@ public class Address implements Parcelable, Visitable<DistrictTypeFactory> {
     }
 
     protected Address(Parcel in) {
-        districtId = in.readInt();
+        districtId = in.readLong();
         districtName = in.readString();
-        cityId = in.readInt();
+        cityId = in.readLong();
         cityName = in.readString();
-        provinceId = in.readInt();
+        provinceId = in.readLong();
         provinceName = in.readString();
         zipCodes = in.createStringArrayList();
     }
@@ -47,11 +47,11 @@ public class Address implements Parcelable, Visitable<DistrictTypeFactory> {
         }
     };
 
-    public int getDistrictId() {
+    public long getDistrictId() {
         return districtId;
     }
 
-    public void setDistrictId(int districtId) {
+    public void setDistrictId(long districtId) {
         this.districtId = districtId;
     }
 
@@ -63,11 +63,11 @@ public class Address implements Parcelable, Visitable<DistrictTypeFactory> {
         this.districtName = districtName;
     }
 
-    public int getCityId() {
+    public long getCityId() {
         return cityId;
     }
 
-    public void setCityId(int cityId) {
+    public void setCityId(long cityId) {
         this.cityId = cityId;
     }
 
@@ -79,11 +79,11 @@ public class Address implements Parcelable, Visitable<DistrictTypeFactory> {
         this.cityName = cityName;
     }
 
-    public int getProvinceId() {
+    public long getProvinceId() {
         return provinceId;
     }
 
-    public void setProvinceId(int provinceId) {
+    public void setProvinceId(long provinceId) {
         this.provinceId = provinceId;
     }
 
@@ -115,11 +115,11 @@ public class Address implements Parcelable, Visitable<DistrictTypeFactory> {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(districtId);
+        dest.writeLong(districtId);
         dest.writeString(districtName);
-        dest.writeInt(cityId);
+        dest.writeLong(cityId);
         dest.writeString(cityName);
-        dest.writeInt(provinceId);
+        dest.writeLong(provinceId);
         dest.writeString(provinceName);
         dest.writeStringList(zipCodes);
         dest.writeList(zipCodes);
